+++ index.html
@@ -230,6 +230,6 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 <script src="https://cdn.jsdelivr.net/npm/lodash@4.17.20/lodash.min.js"></script>
-<script src="/js/main.js" defer></script>
+<script src="/js/main.js" type="module" defer></script>
 </body>
 </html>
